feat(frontend): Create new Layout Component with Register and Login's common properties
