+++ frontend/src/pages/Login.tsx
@@ -1,12 +1,12 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { useSession } from '../hooks/useSession';
 import { LeftArrow } from '../components/svg/LeftArrow';
-import { Box, CssBaseline, IconButton, TextField, ThemeProvider, Typography, createTheme } from '@mui/material';
+import { Box, TextField, Typography } from '@mui/material';
 import { purple } from '../styles/colors';
 import { PurpleButton } from '../components/button/Buttons';
-import { themeSelector } from '../utils/themeSelector';
+import { LoginLayout } from '../components/layout/LoginLayout';
 
 
 export const Login = () => {
@@ -15,53 +15,7 @@ export const Login = () => {
     const navigate = useNavigate();
     const [error, setError] = useState(false);
     const [session] = useSession();
-    const [themeStorage, setThemeStorage] = useState<'light' | 'dark'>('dark');
-
-
-    useEffect(() => {
-        const theme = localStorage.getItem('theme');
-        if (theme) {
-            setThemeStorage(theme as 'light' | 'dark');
-        }
-    }, []);
-
-    const theme = createTheme({
-        palette: {
-            mode: themeStorage,
-        },
-        components: {
-            MuiTextField: {
-                styleOverrides: {
-                    root: {
-                        '& label.Mui-focused': {
-                            color: 'white',
-                        },
-                        '& .MuiOutlinedInput-root': {
-                            fontSize: '0.875rem',
-                            borderRadius: '12px 12px 0px 12px',
-                            '& fieldset': {
-                                borderColor: purple[900],
-                            },
-                            '&:hover fieldset': {
-                                borderColor: purple[700],
-                            },
-                            '&.Mui-focused fieldset': {
-                                borderColor: purple[400],
-                                boxShadow: '0 0 0 3px' + purple[200],
-                                border: '1px solid' + purple[400],
-                            },
-                        },
-                    },
-                }
-            },
-        }
-    });
-
-    const handleThemeChange = (theme: string) => {
-        localStorage.setItem('theme', theme);
-        setThemeStorage(theme as 'light' | 'dark');
-    }
-
+    const [themeStorage, setThemeStorage] = useState<ThemeTypes>('dark');
 
     const onSubmit = (data: any) => {
         // post data to backend
@@ -106,176 +60,115 @@ export const Login = () => {
 
     }
 
-
     return (
-        <>
-            <ThemeProvider theme={theme}>
-                <CssBaseline />
-                <Box
-                    sx={{
-                        position: 'fixed',
-                        top: '0',
-                        left: '0',
-                        right: '0',
-                        display: 'flex',
-                        flexDirection: 'row',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                    }}
-                >
-                    <IconButton onClick={
-                        () => (themeStorage === 'dark' ? handleThemeChange('light') : handleThemeChange('dark'))
-                    } >
-                        {themeSelector[themeStorage]}
-                    </IconButton>
-                </Box>
+        <LoginLayout themeStorage={themeStorage} setThemeStorage={setThemeStorage} maxHeight='400px' title='Login'>
+
+
+            <form onSubmit={
+                handleSubmit(onSubmit)
+            }
+                id='form'
+            >
+
                 <Box
                     sx={{
                         display: 'flex',
                         flexDirection: 'column',
                         alignItems: 'center',
                         justifyContent: 'center',
-                        height: '100vh',
-                        width: '100vw',
+                        gap: 1
                     }}
                 >
-                    <Box
-                        sx={{
-                            display: 'flex',
-                            flexDirection: 'column',
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                            height: '100%',
-                            width: '100%',
-                            maxWidth: '400px',
-                            maxHeight: '400px',
-                            padding: '20px',
-                            borderRadius: '12px',
-                            boxShadow: '0px 0px 10px 0px rgba(0,0,0,0.75)',
-                            backgroundColor: themeStorage === 'dark' ? '#121212' : '#ffffff',
-                        }}
+                    <div className="container-form-userName container-form-input">
+                        <TextField
+                            id="username"
+                            placeholder="Username"
+                            fullWidth
+                            variant="outlined"
+                            sx={{
 
-                    >
-                        <Box>
-                            <Typography
-                                variant='h4'
+                            }}
+                            {...register("username", { required: true })}
+                        />
+                    </div>
+                    {errors.username && <span className="error-msg">This field is required</span>}
+
+                    <div className="container-form-userPassword container-form-input">
+                        <TextField
+                            id="password"
+                            placeholder="Password"
+                            fullWidth
+                            variant="outlined"
+                            type="password"
+                            {...register("password", { required: true })}
+                        />
+
+                    </div>
+                </Box>
+                <Box
+                    sx={{
+                        display: 'flex',
+                        flexDirection: 'row',
+                        alignItems: 'center',
+                        justifyContent: 'center',
+                        gap: 2,
+                        borderRadius: '12px',
+                    }}
+
+                >
+                    <PurpleButton
+                        variant='contained'
+                        width='75%'
+                        children={
+                            <Box
                                 sx={{
-                                    color: themeStorage === 'dark' ? 'white' : 'black',
-                                    fontWeight: 'bold',
-                                    textAlign: 'center',
-                                    mb: 5,
+                                    display: 'flex',
+                                    justifyContent: 'center',
+                                    alignItems: 'center',
+                                }}
+                            >
+                                Register
+                            </Box>
+                        }
+                        color='white'
+                        margin='10px 0px 10px 0px'
+                        onClick={() => {
+                            navigate('/register');
+                        }}
+                        backgroundColor={purple[800]}
+                    />
+                    <PurpleButton
+                        variant='contained'
+                        width='75%'
+                        color='white'
+                        children={
+                            <Box
+                                sx={{
+                                    display: 'flex',
+                                    justifyContent: 'center',
+                                    alignItems: 'center',
                                 }}
                             >
                                 Login
-                            </Typography>
-                        </Box>
-                        <form onSubmit={
-                            handleSubmit(onSubmit)
+                            </Box>
                         }
-                            id='form'
-                        >
-
-                            <Box
-                                sx={{
-                                    display: 'flex',
-                                    flexDirection: 'column',
-                                    alignItems: 'center',
-                                    justifyContent: 'center',
-                                    gap: 1
-                                }}
-                            >
-                                <div className="container-form-userName container-form-input">
-                                    <TextField
-                                        id="username"
-                                        placeholder="Username"
-                                        fullWidth
-                                        variant="outlined"
-                                        {...register("username", { required: true })}
-                                    />
-                                </div>
-                                {errors.username && <span className="error-msg">This field is required</span>}
-
-                                <div className="container-form-userPassword container-form-input">
-                                    <TextField
-                                        id="password"
-                                        placeholder="Password"
-                                        fullWidth
-                                        variant="outlined"
-                                        type="password"
-                                        {...register("password", { required: true })}
-                                    />
-
-                                </div>
-                            </Box>
-                            <Box
-                                sx={{
-                                    display: 'flex',
-                                    flexDirection: 'row',
-                                    alignItems: 'center',
-                                    justifyContent: 'center',
-                                    gap: 2,
-                                    borderRadius: '12px',
-                                }}
-
-                            >
-                                <PurpleButton
-                                    variant='contained'
-                                    width='75%'
-                                    children={
-                                        <Box
-                                            sx={{
-                                                display: 'flex',
-                                                justifyContent: 'center',
-                                                alignItems: 'center',
-                                            }}
-                                        >
-                                            Register
-                                        </Box>
-                                    }
-                                    color='white'
-                                    margin='10px 0px 10px 0px'
-                                    onClick={() => {
-                                        navigate('/register');
-                                    }}
-                                    backgroundColor={purple[800]}
-                                />
-                                <PurpleButton
-                                    variant='contained'
-                                    width='75%'
-                                    color='white'
-                                    children={
-                                        <Box
-                                            sx={{
-                                                display: 'flex',
-                                                justifyContent: 'center',
-                                                alignItems: 'center',
-                                            }}
-                                        >
-                                            Login
-                                        </Box>
-                                    }
-                                    type='submit'
-                                    margin='10px 0px 10px 0px'
-                                    backgroundColor={purple[800]}
-                                />
-                            </Box>
-                            {
-                                error && <span className="error-msg">Invalid username or password</span>
-                            }
-
-                            <div className="info-msg-container">
-                                {
-                                    session && <span className="info-msg">
-                                        <a href="/"><LeftArrow height='24px' width='24px' /></a>
-                                        You are already logged in</span>
-                                }
-                            </div>
-                        </form>
-
-
-                    </Box>
+                        type='submit'
+                        margin='10px 0px 10px 0px'
+                        backgroundColor={purple[800]}
+                    />
                 </Box>
-            </ThemeProvider>
-        </>
+                {
+                    error && <span className="error-msg">Invalid username or password</span>
+                }
+
+                <div className="info-msg-container">
+                    {
+                        session && <span className="info-msg">
+                            <a href="/"><LeftArrow height='24px' width='24px' /></a>
+                            You are already logged in</span>
+                    }
+                </div>
+            </form>
+        </LoginLayout>
     );
 }
